+++ package.json
@@ -14,6 +14,7 @@
     "bluebird": "^3.5.1",
     "discord.js": "^11.3.2",
     "npm": "^6.1.0",
-    "tedious": "^2.6.1"
+    "tedious": "^2.6.1",
+    "util": "^0.10.0"
   }
 }
